Add deployment to travis script

+++ deploy-docs.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 set -e # exit with nonzero exit code if anything fails
 
+# Travis may run multiple jobs, we should only deploy on the deploy version
+if [ "${DEPLOY_VERSION}" == "${RACKET_VERSION}" ]; then exit 0; fi
+
 # clear and re-create the out directory
 rm -rf out || exit 0;
 mkdir out;
